% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_cos_similarity.R
\name{avg_cos_similarity}
\alias{avg_cos_similarity}
\title{Computes the average similarity vector between a set of cues and the full vocab over multiple embedding models}
\usage{
avg_cos_similarity(embeds_list, cues, method = "cosine", norm = "l2")
}
\arguments{
\item{embeds_list}{list of trained word embedding models (each is a V by D matrix)}

\item{cues}{a character vector of cue words for which the similarity vector will be computed}

\item{method}{a \code{param} of \code{text2vec}'s \code{sim2}: character, the similarity measure to be used. One of c("cosine","jaccard").}

\item{norm}{a \code{param} of \code{text2vec}'s \code{sim2}: character = c("l2","none") - how to scale input matrices. If they already scaled - use "none"}
}
\value{
a matrix of similarity vectors, one row for each cue
}
\description{
Computes the average similarity vector between a set of cues and the full vocab over multiple embedding models
}
\note{
embedding models should not be averaged directly, instead, users should average
the similarity vectors. This function makes this easier. See Rodriguez & Spirling.
}
