% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_embeds.R
\name{corr_embeds}
\alias{corr_embeds}
\title{Compute the correlation of a given cue's similarity vectors for two embedding models}
\usage{
corr_embeds(
  embeds1,
  embeds2,
  cue,
  type = "pearson",
  method = "cosine",
  norm = "l2"
)
}
\arguments{
\item{embeds1}{embeddings model 1 (V by D matrix)}

\item{embeds2}{embeddings model 2 (V by D matrix)}

\item{cue}{cue word for which the correlation will be computed}

\item{type}{\code{pearson} or \code{rank} correlation}

\item{method}{a \code{param} of \code{text2vec}'s \code{sim2}: character, the similarity measure to be used. One of c("cosine","jaccard").}

\item{norm}{a \code{param} of \code{text2vec}'s \code{sim2}: character = c("l2","none") - how to scale input matrices. If they already scaled - use "none"}
}
\value{
a numeric value between -1 and 1
}
\description{
Compute the correlation of a given cue's similarity vectors for two embedding models
}
\note{
embedding models will be subsetted to common vocabulary
}
