% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cue_sim.R
\name{cue_sim}
\alias{cue_sim}
\title{Compute similarity vector between a cue word and all words in the vocabulary of an embeddings model}
\usage{
cue_sim(cue, embeds, method = "cosine", norm = "l2", rank = FALSE)
}
\arguments{
\item{cue}{cue word for which similarity vector will be computed}

\item{embeds}{embeddings model (V by D matrix)}

\item{method}{a \code{param} of \code{text2vec}'s \code{sim2}: character, the similarity measure to be used. One of c("cosine","jaccard").}

\item{norm}{a \code{param} of \code{text2vec}'s \code{sim2}: character = c("l2","none") - how to scale input matrices. If they already scaled - use "none"}

\item{rank}{logical value indicating whether to return a vector of similarities or ranks}
}
\value{
a numeric vector of cosine similarities (or ranks if indicated) between all words in the vocabulary and the cue word
}
\description{
Compute similarity vector between a cue word and all words in the vocabulary of an embeddings model
}
