% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_sims.R
\name{jaccard_sims}
\alias{jaccard_sims}
\title{Compute jaccard index of two embeddings models for N nearest neighbors}
\usage{
jaccard_sims(sims1, sims2, N = 10, common_vocab = FALSE)
}
\arguments{
\item{sims1}{matrix of similarities (e.g. cosine) between a set of cues (rows) and the full vocab (columns)}

\item{sims2}{matrix of similarities (e.g. cosine) between a set of cues (rows) and the full vocab (columns)}

\item{N}{number of nearest neighbors with which to compute jaccard index}

\item{common_vocab}{whether to use a common vocabulary to compute jaccard index}
}
\value{
a tibble with mean and standard error of jaccard index overt set of cues
}
\description{
Compute jaccard index of two embeddings models for N nearest neighbors
}
