% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_neighbors.R
\name{nearest_neighbors}
\alias{nearest_neighbors}
\title{Find the nearest neighbors of a cue given a trained word embedding model}
\usage{
nearest_neighbors(cue, embeds, N = 5, method = "cosine", norm = "l2")
}
\arguments{
\item{cue}{cue word for which we want nearest neighbors}

\item{embeds}{matrix of trained embeddings}

\item{N}{number of nearest neighbors to return}

\item{method}{a \code{param} of \code{text2vec}'s \code{sim2}: character, the similarity measure to be used. One of c("cosine","jaccard").}

\item{norm}{a \code{param} of \code{text2vec}'s \code{sim2}: character = c("l2","none") - how to scale input matrices. If they already scaled - use "none"}
}
\value{
a character vector of nearest neighbors of length N
}
\description{
Find the nearest neighbors of a cue given a trained word embedding model
}
\note{
\code{context} the first nearest neighbor will be the cue itself.
}
