% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_sims.R
\name{nn_sims}
\alias{nn_sims}
\title{Return nearest neighbors based on cosine similarity vector}
\usage{
nn_sims(cue, sims, N = 5, drop_cue = TRUE)
}
\arguments{
\item{cue}{cue word for which nearest neighbors are required}

\item{sims}{matrix of similariteis (with cues as rownames)}

\item{N}{number of nearest neighbors to return}
}
\value{
a character vector of nearest neighbors to cue word
}
\description{
Return nearest neighbors based on cosine similarity vector
}
\note{
cue is always the nearest neighbor hence can request to drop
}
